+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Added more tracing for input focus handling when using the `+editor`
+  `YABRIDGE_DEBUG_LEVEL` flag.
+
 ## [3.5.0] - 2021-06-23
 
 ### Added